refactor(memory_ipc): split into mod/server/client (#1266)
Promote memory_ipc.rs (785 LOC) to dir module:
- mod.rs (332): wire types + helpers + SharedMemory + reap_stale_open_lock
- server.rs (215): spawn_server + ServerHandle + serve_connection + dispatch
- client.rs (272): RemoteCognitiveMemory + CognitiveMemoryOps impl

All 6 memory_ipc tests pass.
